// src/features/clientes/types.ts
export type Cliente = {
    id: number
    nome: string
    telefone: string
    obs: string | null
    cep: string | null
    logradouro: string | null
    numero: string | null
    complemento: string | null
    bairro: string | null
    cidade: string | null
    uf: string | null
    createdAt: string
}